Recombine files so Pull request diff is not ridiculous
