--- src/execution/context.js
@@ -1,88 +1,0 @@
-/* @flow */
-/**
- *  Copyright (c) 2015, Facebook, Inc.
- *  All rights reserved.
- *
- *  This source code is licensed under the BSD-style license found in the
- *  LICENSE file in the root directory of this source tree. An additional grant
- *  of patent rights can be found in the PATENTS file in the same directory.
- */
-import { GraphQLError } from '../error';
-import { GraphQLSchema } from '../type/schema';
-import type {
-  OperationDefinition,
-  Document,
-  FragmentDefinition
-} from '../language/ast';
-import { getVariableValues } from './values';
-import { Kind } from '../language';
-
-/**
- * Data that must be available at all points during query execution.
- *
- * Namely, schema of the type system that is currently executing,
- * and the fragments defined in the query document
- */
-export type ExecutionContext = {
-  schema: GraphQLSchema;
-  fragments: {[key: string]: FragmentDefinition};
-  rootValue: mixed;
-  operation: OperationDefinition;
-  variableValues: {[key: string]: mixed};
-  errors: Array<GraphQLError>;
-}
-
-/**
- * Constructs a ExecutionContext object from the arguments passed to
- * execute, which we will pass throughout the other execution methods.
- *
- * Throws a GraphQLError if a valid execution context cannot be created.
- */
-export function buildExecutionContext(
-  schema: GraphQLSchema,
-  documentAST: Document,
-  rootValue: mixed,
-  rawVariableValues: ?{[key: string]: mixed},
-  operationName: ?string
-): ExecutionContext {
-  const errors: Array<GraphQLError> = [];
-  let operation: ?OperationDefinition;
-  const fragments: {[name: string]: FragmentDefinition} = Object.create(null);
-  documentAST.definitions.forEach(definition => {
-    switch (definition.kind) {
-      case Kind.OPERATION_DEFINITION:
-        if (!operationName && operation) {
-          throw new GraphQLError(
-            'Must provide operation name if query contains multiple operations.'
-          );
-        }
-        if (!operationName ||
-            definition.name && definition.name.value === operationName) {
-          operation = definition;
-        }
-        break;
-      case Kind.FRAGMENT_DEFINITION:
-        fragments[definition.name.value] = definition;
-        break;
-      default: throw new GraphQLError(
-        `GraphQL cannot execute a request containing a ${definition.kind}.`,
-        definition
-      );
-    }
-  });
-  if (!operation) {
-    if (!operationName) {
-      throw new GraphQLError(`Unknown operation named "${operationName}".`);
-    } else {
-      throw new GraphQLError('Must provide an operation.');
-    }
-  }
-  const variableValues = getVariableValues(
-    schema,
-    operation.variableDefinitions || [],
-    rawVariableValues || {}
-  );
-  const exeContext: ExecutionContext =
-    { schema, fragments, rootValue, operation, variableValues, errors };
-  return exeContext;
-}
