+++ src/execution/execute.js
@@ -22,6 +22,8 @@
 import { GraphQLError, locatedError } from '../error';
 import invariant from '../jsutils/invariant';
 import isNullish from '../jsutils/isNullish';
+import { Kind } from '../language';
+import { getVariableValues } from './values';
 
 import {
   GraphQLScalarType,
@@ -40,12 +42,10 @@ import type {
 } from '../type/definition';
 import { GraphQLSchema } from '../type/schema';
 import type {
-  Document
+  OperationDefinition,
+  Document,
+  FragmentDefinition
 } from '../language/ast';
-import {
-  ExecutionContext,
-  buildExecutionContext
-} from './context';
 import {
   planOperation
 } from './plan';
@@ -95,6 +95,21 @@ import {
  * | completeValue    | planSelection     | GraphQLSelectionCompletionPlan     |
  * +------------------+-------------------+------------------------------------+
  */
+
+/**
+ * Data that must be available at all points during query execution.
+ *
+ * Namely, schema of the type system that is currently executing,
+ * and the fragments defined in the query document
+ */
+export type ExecutionContext = {
+  schema: GraphQLSchema;
+  fragments: {[key: string]: FragmentDefinition};
+  rootValue: mixed;
+  operation: OperationDefinition;
+  variableValues: {[key: string]: mixed};
+  errors: Array<GraphQLError>;
+}
 
 /**
  * The result of execution. `data` is the result of executing the
@@ -163,6 +178,61 @@ export function execute(
     }
     return { data, errors: context.errors };
   });
+}
+
+/**
+ * Constructs a ExecutionContext object from the arguments passed to
+ * execute, which we will pass throughout the other execution methods.
+ *
+ * Throws a GraphQLError if a valid execution context cannot be created.
+ */
+export function buildExecutionContext(
+  schema: GraphQLSchema,
+  documentAST: Document,
+  rootValue: mixed,
+  rawVariableValues: ?{[key: string]: mixed},
+  operationName: ?string
+): ExecutionContext {
+  const errors: Array<GraphQLError> = [];
+  let operation: ?OperationDefinition;
+  const fragments: {[name: string]: FragmentDefinition} = Object.create(null);
+  documentAST.definitions.forEach(definition => {
+    switch (definition.kind) {
+      case Kind.OPERATION_DEFINITION:
+        if (!operationName && operation) {
+          throw new GraphQLError(
+            'Must provide operation name if query contains multiple operations.'
+          );
+        }
+        if (!operationName ||
+            definition.name && definition.name.value === operationName) {
+          operation = definition;
+        }
+        break;
+      case Kind.FRAGMENT_DEFINITION:
+        fragments[definition.name.value] = definition;
+        break;
+      default: throw new GraphQLError(
+        `GraphQL cannot execute a request containing a ${definition.kind}.`,
+        definition
+      );
+    }
+  });
+  if (!operation) {
+    if (!operationName) {
+      throw new GraphQLError(`Unknown operation named "${operationName}".`);
+    } else {
+      throw new GraphQLError('Must provide an operation.');
+    }
+  }
+  const variableValues = getVariableValues(
+    schema,
+    operation.variableDefinitions || [],
+    rawVariableValues || {}
+  );
+  const exeContext: ExecutionContext =
+    { schema, fragments, rootValue, operation, variableValues, errors };
+  return exeContext;
 }
 
 /**
